requirements.txt
python-telegram-bot==21.5
google-api-python-client>=2.111.0
google-auth-oauthlib
google-auth-httplib2
python-dotenv
